requirements.txt
asyncpg==0.30.0
fastapi==0.115.8
pydantic==2.10.6
python_bcrypt==0.3.2
